RONDB-914 Increase support of API nodes from 255 to 2039
This patch takes the first step towards supporting more API nodes.
We replace MAX_NODES and MAX_NDB_NODES by calls to functions that
return the max number in the current cluster. We introduce
ABS_MAX_NODES and ABS_MAX_NDB_NODES for those cases where we need
to keep the exact limit.

This solves most of the problem. In addition the signal SUB_TABLE_DATA
sent a data node id (8 bits), an API node id (8 bits) and a request info
of 8 bits. The first fix is to increase API node id to 16 bits.

The second problem is schema transactions that use a table called
ndb_schema. This kept a bit map in a BINARY(32) column. We replaced this
by a BINARY(255) which means that the limit is 2039, we can have node ids
from 1 to 2039.

This patch also fixes a problem in that TCKEY_FAILCONF didn't send a
TC_COMMIT_ACK in cases where the API already released the NdbReceiver
object.

Changed test cases such that a number of them used node ids beyond 255
in many cases.

This patch makes it possible to start a new cluster with more node ids.
It doesn't handle upgrade and downgrade functionality.

When a node retrieves a configuration it will provide its version in
this version if it wants to use a configuration with a node id above
255, otherwise the get config request will be refused.

When an API node registers with a high node id we will check that
all nodes that are connected to the data node can handle the high
node id.

If an API node or a data node registers in a data node and that
data node have a configuration with high node ids, it will require
all nodes to register to have the version that supports high
node ids.

The ndb_schema table has extended the length of the slock column
to 255 bytes, thus leading to supporting 2039 node ids from 1
to 2039. In an upgrade situation this means that this table needs
to be dropped and recreated if the ndb_schema table only supports
node ids from 1 to 255.

+++ storage/ndb/src/kernel/blocks/dbtup/DbtupBuffer.cpp
@@ -54,7 +54,7 @@ void Dbtup::execSEND_PACKED(Signal *signal) {
   for (i = 0; i < TpackedListIndex; i++) {
     jam();
     hostId = cpackedList[i];
-    ndbrequire((hostId - 1) < (MAX_NODES - 1));  // Also check not zero
+    ndbrequire(Uint32((hostId - 1)) < Uint32((MAX_NODES - 1)));  // Also check not zero
     HostBuffer *const buffer = &hostBuffer[hostId];
     Uint32 TpacketTA = buffer->noOfPacketsTA;
     if (TpacketTA != 0) {
